[Exceptions] #4 Did some rework on the exceptions since we need more exception types right now. Added a interface for the advice classes (BaseAdvice) so all advice classes will have a handler function.
Theoretically we could unify the advice classes with a default advice class (that implements the interface) and extend that such that all advice classes have the same default behavior which will be in the default (parent) class.

This will result in a more streamlined exception build up and we could even add a factory that can create the diffrent exception classes. But for now this will do. Gotta focus on the actual requirements first and maybe when there is time left we can refactor the whole exception structure.