[Customer service] #5 Adding logic to the CustomerService and changing some return types in the BaseService.
Changed return type for update and create method since we want some confirmation on the other side (client-side) when customer data is saved succesfully. Let's say we want to update the client-side with the new data that we just saved or created. We should therefor send back the entity that is created by hibernate as a dto.