[Customer controller] #3 Finishing up the customer controller. Functionality is implemented as the requirements ... required.
On to the next controller.
When working with a team bigger then one person, it would be wise to start branching. This way someone can work on the customer controller while another team member can work on another part without getting in each other's way. After finishing development and testing of the functionality the branch can be merged back to the master/main branch.