[Transaction Service] #7 Implementing the needed methods and functionality for the `TransactionService` and `TransactionRepository`.
~ Transaction:
added enum types for the transaction type so we know is we should subtract of add to an account whenever we start doing the transactions (whenever there is time left)

~ TransactionDto:
added missing @Data annotation. Which made Spring panic.

~ TransactionRepository:
added custom method `findAllByAccount()` so we are able to retrieve all transactions for the given account.

~ TransactionService:
implemented `getAllById()` method which will return all transactions for the given account, using the repo method we added earlier.