[AccountController/Service] #2 #6 Working on the controller and service so we can create a new account for customerId.
Added custom modelMapper configs to circumvent the final variables on the entities. The default mapper does not set these since the default mapper tries to set data to the object using the setters.

A lot of debugging.

~ AccountController:
This is still a WIP. Been working on the /create endpoint. Accounts can be created and we not just need to creation of Transactions to work. Since it is late now and all i will do this tommorow. Focus is kinda off now after a day of work and this api ;)

~ AccountDto:
Changed datatype of variable type to Account.AccountTypes to ensure the datatype.

~ AccountService:
added custom mapper config so dto's get mapped to entities correctly. The default config cannot map final fields and this circumvents the problem.

reworked the generateNewAccountNumber method since that was dumb and borked. It now currectly generates an account number with a max of 18 characters.