[AccountController & Service] #2 #6 finishing the creation of an account and the first transaction is the initial credit is greater then zero.
~ AccountController:
Added documentation and moved the account creation up so we have an reference when creating the transaction, since that needs the accountId.

~ AccountService:
Added logic to the get method to return the queried account with all the transactions. Modified the getAll() method so it also returns the transactions for each account.